Added comments to use_python code and renamed python wrapper to py_astro_accelerate.py
